mm-hotplug-verify-hotplug-memory-range-fix
avoid printk warnings

Cc: KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>
Cc: Tang Chen <tangchen@cn.fujitsu.com>
Cc: Toshi Kani <toshi.kani@hp.com>
Cc: Wanpeng Li <liwanp@linux.vnet.ibm.com>

+++ mm/memory_hotplug.c
@@ -1079,7 +1079,8 @@ static int check_hotplug_memory_range(u64 start, u64 size)
 	if ((start_pfn & ~PAGE_SECTION_MASK) ||
 	    (nr_pages % PAGES_PER_SECTION) || (!nr_pages)) {
 		pr_err("Section-unaligned hotplug range: start 0x%llx, size 0x%llx\n",
-				start, size);
+				(unsigned long long)start,
+				(unsigned long long)size);
 		return -EINVAL;
 	}
